guardado de atestiguamiento

+++ source/dictamen/dictamen-soa/dictamen-soa-commons/src/main/java/mx/gob/imss/cit/dictamen/commons/to/domain/PreguntaTO.java
@@ -17,7 +17,7 @@ public class PreguntaTO extends BaseTO {
     private RubroTO cveIdRubro;
     private PreguntaTO cveIdPreguntaPadre;
     private List<OpcionPreguntaTO> ndcOpcionesPregunta;
-    private Integer opcionSeleccionada;
+    private Long opcionSeleccionada;
     private String observaciones;
     
     
@@ -75,10 +75,10 @@ public class PreguntaTO extends BaseTO {
 	public void setNdcOpcionesPregunta(List<OpcionPreguntaTO> ndcOpcionesPregunta) {
 		this.ndcOpcionesPregunta = ndcOpcionesPregunta;
 	}
-	public Integer getOpcionSeleccionada() {
+	public Long getOpcionSeleccionada() {
 		return opcionSeleccionada;
 	}
-	public void setOpcionSeleccionada(Integer opcionSeleccionada) {
+	public void setOpcionSeleccionada(Long opcionSeleccionada) {
 		this.opcionSeleccionada = opcionSeleccionada;
 	}
 	public String getObservaciones() {
